10.7a initial build for migrating permanently to 10.7
10.7a initial build for migrating permanently to 10.7

+++ Classes/Utils.m
@@ -67,7 +67,7 @@ static Utils* sharedInstance = nil;
     return self;
 }
 
-- (NSUInteger) retainCount {
+- (unsigned long) retainCount {
     return UINT_MAX;  // denotes an object that cannot be released
 }
